+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [v1.0.0] - 07/22/2024
+- Remove unnecessary domain enumeration functions
+- Remove domain enumeration flags from CLI
+- Update install-tools script to properly install all dependencies on Debian 12
+
 ## [v0.4.2] - 05/24/2024
 - Remove unnecessary and bloated .cache folder generated from vulnerability scans and default logins scans.
 
@@ -9,7 +14,6 @@
 - Add ability to provide exclude file to full port scan operation
 
 ## [v0.3.0] - 04/24/2024
-
 - Remove service enumeration checks from full port nmap scans
 
 ## [v0.2.0] - 04/22/2024